Begining to add player animations
The addtion of code to acess the spritesheets. This will be implimented for 8 possible animations of the ship (Engine: idle, Off, Up, Down, Boosted Up, Boosted Down, Boosted Forward; and Blown up). There will need to be a linked up to the spritesheet to account for each motion of the ship that is possible.